docs :: common enum 분리
