+++ apps/admin/src/@types/enums/common.ts
@@ -1,0 +1,6 @@
+export const CompanyTypeEnToKr = {
+  LEAD: "선도기업",
+  PARTICIPATING: "참여기업",
+  CONTRACTING: "협약기업",
+  DEFAULT: "기본",
+} as const;
