+++ apps/admin/src/@types/enums/index.ts
@@ -1,1 +1,2 @@
+export * from "./common";
 export * from "./recruitment";
